Integrated Dyn4J for Transform management

+++ src/main/java/eu/saltyscout/leo2d/render/particle/FloatArrayCurve.java
@@ -5,15 +5,15 @@ import eu.saltyscout.leo2d.util.MathUtil;
 /**
  * Created by Peter on 01.11.2017.
  */
-public class FloatArrayCurve extends Curve<float[]> {
+public class FloatArrayCurve extends Curve<double[]> {
 
-    public FloatArrayCurve(float[] start) {
+    public FloatArrayCurve(double[] start) {
         super(start);
     }
 
     @Override
-    float[] lerp(float pc, float[] a, float[] b) {
-        float[] r = new float[a.length];
+    double[] lerp(float pc, double[] a, double[] b) {
+        double[] r = new double[a.length];
         for (int i = 0; i < r.length; i++) {
             r[i] = MathUtil.lerp(pc, a[i], b[i]);
         }
